finition matchscore et début matchduel

+++ sae_java/src/main/java/EquipeScore.java
@@ -1,7 +1,7 @@
 public class EquipeScore<X, Y> { 
     public final Equipe equipe; 
-    public final Integer score; 
-    public EquipeScore(Equipe equipe, Integer score) { 
+    public final Double score; 
+    public EquipeScore(Equipe equipe, Double score) { 
       this.equipe = equipe; 
       this.score = score; 
     }
